+++ python/pyqrlew/io/postgresql.py
@@ -3,7 +3,8 @@ from pathlib import Path
 import importlib.resources as pkg_resources
 import json
 import logging
-from sqlalchemy import MetaData, Engine, exists, select, column, table, text, Result
+from sqlalchemy import MetaData, exists, select, column, table, text
+# from sqlalchemy.engine import Engine, Result
 from datasets import sources
 from datasets.databases import PostgreSQL as EmptyPostgreSQL
 from pyqrlew.io.database import dataset