+++ src/dataset.rs
@@ -1,4 +1,4 @@
-
+use std::ops::Deref;
 use qrlew::{builder::With, sql};
 use qrlew_sarus::data_spec;
 use std::rc::Rc;
@@ -7,7 +7,22 @@ use pyo3::{pyclass, pymethods};
 use crate::{relation::Relation, error::Result};
 
 #[pyclass]
-pub struct Dataset(pub data_spec::Dataset);
+#[derive(Clone)]
+pub struct Dataset(data_spec::Dataset);
+
+impl Deref for Dataset {
+    type Target = data_spec::Dataset;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl From<Dataset> for data_spec::Dataset {
+    fn from(value: Dataset) -> Self {
+        value.0
+    }
+}
 
 #[pymethods]
 impl Dataset {
@@ -19,25 +34,20 @@ impl Dataset {
     }
 
     pub fn relations(&self) -> Vec<(Vec<String>, Relation)> {
-        self.0
-            .relations()
+        self.deref().relations()
             .into_iter()
-            .map(|(i, r)| (i, Relation(r)))
+            .map(|(i, r)| (i, Relation::new(r)))
             .collect()
     }
 
     pub fn sql(&self, query: &str) -> Result<Relation> {
         let query = sql::relation::parse(query)?;
-        let relations = self.0.relations();
+        let relations = self.deref().relations();
         let query_with_relations = query.with(&relations);
-        Ok(Relation(Rc::new(query_with_relations.try_into()?)))
+        Ok(Relation::new(Rc::new(query_with_relations.try_into()?)))
     }
 
     pub fn __str__(&self) -> String {
         format!("{}", self.0)
     }
-
-    // fn schema_type_data(&self) -> String {
-    //     (self.0).schema_type_data()
-    // }
 }