+++ src/relation.rs
@@ -1,28 +1,49 @@
 use pyo3::prelude::*;
-use qrlew::relation::Variant;
-use qrlew::{relation, ast};
+use qrlew::{
+    ast,
+    data_type::DataType,
+    relation::{self, Variant},
+    differential_privacy::DPRelation,
+    protection::PEPRelation};
 use serde_json::Value;
-use std::rc::Rc;
+use std::{rc::Rc, ops::Deref};
 use crate::{error::Result, dataset::Dataset};
-
+use qrlew_sarus::protobuf::{type_, schema, print_to_string};
+use std::str;
 
 #[pyclass(unsendable)]
-pub struct Relation(pub Rc<relation::Relation>);
+#[derive(Clone)]
+pub struct Relation(Rc<relation::Relation>);
+
+impl Deref for Relation {
+    type Target = relation::Relation;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl Relation {
+    pub fn new(relation: Rc<relation::Relation>) -> Self {
+        Relation(relation)
+    }
+}
 
 #[pymethods]
 impl Relation {
-    pub fn parse(&self, query: &str, dataset: &Dataset) -> Result<Self> {
+    #[staticmethod]
+    pub fn parse(query: &str, dataset: &Dataset) -> Result<Self> {
         dataset.sql(query)
     }
 
     pub fn __str__(&self) -> String {
-        let relation = &*(self.0);
+        let relation = self.0.as_ref();
         format!("{}", relation)
     }
 
     pub fn dot(&self) -> Result<String> {
         let mut out: Vec<u8> = vec![];
-        (*self.0).dot(&mut out, &[]).unwrap();
+        self.0.as_ref().dot(&mut out, &[]).unwrap();
         Ok(String::from_utf8(out).unwrap())
     }
 
@@ -30,122 +51,21 @@ impl Relation {
         (*self.0).schema().to_string()
     }
 
-    pub fn protect(&self, dataset: &Dataset, protected_entity: &str) -> Result<Self> {
-        let pe = parse_protected_entity(protected_entity);
-        Ok(Relation(Rc::new(protect((*(self.0)).clone(), dataset, &pe)?)))
+    pub fn protect<'a>(&'a self, dataset: &'a Dataset, protected_entity: Vec<(&'a str, Vec<(&'a str, &'a str, &'a str)>, &'a str)>) -> Result<Self> {
+        let relations = dataset.deref().relations();
+        Ok(Relation(Rc::new(self.deref().clone().force_protect_from_field_paths(&relations, protected_entity).into())))
     }
 
-    pub fn dp_compilation(&self, dataset: &Dataset, protected_entity: &str, epsilon: f64, delta: f64) -> Result<Self> {
-        let pe = parse_protected_entity(protected_entity);
-        Ok(Relation(Rc::new(dp_compilation((*(self.0)).clone(), dataset, &pe, epsilon, delta)?)))
+    pub fn dp_compile<'a>(&'a self, dataset: &'a Dataset, protected_entity: Vec<(&'a str, Vec<(&'a str, &'a str, &'a str)>, &'a str)>, epsilon: f64, delta: f64) -> Result<Self> {
+        let relations = dataset.deref().relations();
+        let pep_relation = self.deref().clone().force_protect_from_field_paths(&relations, protected_entity);
+        let dp_relation = pep_relation.dp_compile(epsilon, delta)?;
+        Ok(Relation(Rc::new(dp_relation.into())))
     }
 
     pub fn render(&self) -> String {
         let relation = &*(self.0);
         let ast_query: ast::Query = relation.into();
         format!("{}", ast_query)
-    }
-}
-
-fn protect<'a>(
-    relation: relation::Relation,
-    dataset: &Dataset,
-    vec_of_string: &'a Vec<(String, Vec<(String, String, String)>, String)>,
-) -> Result<relation::Relation> {
-    let relations = dataset.0.relations();
-    let vec_pe = vec_of_string
-        .iter()
-        .map(|(t, fk, n)| {
-            let bidings = fk.iter()
-            .map(|(a1, a2, a3)| (a1.as_str(), a2.as_str(), a3.as_str()))
-            .collect::<Vec<_>>();
-            (
-                t.as_str(),
-                bidings,
-                n.as_str(),
-            )
-        }).collect::<Vec<_>>();
-    let slice_pe = vec_pe.iter()
-    .map(|(t, fk, n)| (*t, fk.as_slice(), *n)).collect::<Vec<_>>();
-    Ok(relation.force_protect_from_field_paths(&relations, slice_pe.as_slice()))
-}
-
-fn dp_compilation<'a>(
-    relation: relation::Relation,
-    dataset: &Dataset,
-    vec_of_string: &'a Vec<(String, Vec<(String, String, String)>, String)>,
-    epsilon: f64,
-    delta: f64
-) -> Result<relation::Relation> {
-    let relations = dataset.0.relations();
-    let vec_pe = vec_of_string
-        .iter()
-        .map(|(t, fk, n)| {
-            let bidings = fk.iter()
-            .map(|(a1, a2, a3)| (a1.as_str(), a2.as_str(), a3.as_str()))
-            .collect::<Vec<_>>();
-            (
-                t.as_str(),
-                bidings,
-                n.as_str(),
-            )
-        }).collect::<Vec<_>>();
-    let slice_pe = vec_pe.iter()
-    .map(|(t, fk, n)| (*t, fk.as_slice(), *n)).collect::<Vec<_>>();
-    Ok(relation.dp_compilation(&relations, slice_pe.as_slice(), epsilon, delta)?)
-}
-
-fn parse_protected_entity(str_pe: &str) -> Vec<(String, Vec<(String, String, String)>, String)> {
-    let v: Value = serde_json::from_str(str_pe).expect("Failed to parse JSON");
-    let mut peid: Vec<(String, Vec<(String, String, String)>, String)> = vec![];
-
-    if let Value::Array(json_array) = &v["protected_entity"] {
-        for table in json_array {
-            if let Value::Array(table_array) = &table {
-                assert_eq!(table_array.len(), 3);
-                let tablename = if let Value::String(s) = &table_array[0] {
-                    s.to_string()
-                } else {
-                    panic!()
-                };
-                let name = if let Value::String(s) = &table_array[2] {
-                    s.to_string()
-                } else {
-                    panic!()
-                };
-                let foreign_keys: Vec<(String, String, String)> =
-                    if let Value::Array(a) = &table_array[1] {
-                        a.iter()
-                            .map(|v| {
-                                if let Value::Array(a) = &v {
-                                    assert_eq!(a.len(), 3);
-                                    let my_list = a
-                                        .iter()
-                                        .map(|e| {
-                                            if let Value::String(s) = e {
-                                                s.to_string()
-                                            } else {
-                                                panic!()
-                                            }
-                                        })
-                                        .collect::<Vec<String>>();
-                                    (my_list[0].clone(), my_list[1].clone(), my_list[2].clone())
-                                } else {
-                                    println!("v = {:?}", v);
-                                    panic!()
-                                }
-                            })
-                            .collect()
-                    } else {
-                        panic!()
-                    };
-                peid.push((tablename, foreign_keys, name))
-            } else {
-                panic!()
-            }
-        }
-        peid
-    } else {
-        panic!()
     }
 }