+++ tests/test_extract_dataset.py
@@ -1,4 +1,5 @@
-from sqlalchemy import Engine, Connection, text
+from sqlalchemy import text
+from sqlalchemy.engine import Connection
 import typing as t
 import os
 from termcolor import colored