+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- update qrlew version to 0.3 [MR15](https://github.com/Qrlew/pyqrlew/pull/15)
+- downgrade sqlalchemy to 1.4.* [MR15](https://github.com/Qrlew/pyqrlew/pull/15)
+- downgrade to psycopg2 [MR15](https://github.com/Qrlew/pyqrlew/pull/15)
 
 ## [0.3.3] - 2023-08-29
 ### Changed