+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.6] - 2023-09-29
+### Changed
+- tau_thresholding epsilon and delta can be set
+
+## [0.3.5] - 2023-09-29
+### Changed
+- Updated qrlew where objects are threadsafe now
+
+## [0.3.4] - 2023-09-28
 ### Changed
 - update qrlew version to 0.3 [MR15](https://github.com/Qrlew/pyqrlew/pull/15)
 - downgrade sqlalchemy to 1.4.* [MR15](https://github.com/Qrlew/pyqrlew/pull/15)